tests: Move utils.js out of polyfill folder
