+++ web/packages/selfhosted/test/.eslintrc.json
