--- web/packages/selfhosted/test/polyfill/.eslintrc.json
@@ -1,8 +1,0 @@
-{
-    "env": {
-        "mocha": true
-    },
-    "globals": {
-        "browser": true
-    }
-}
