+++ web/packages/selfhosted/test/polyfill/pdf_with_get/test.js
@@ -1,4 +1,4 @@
-const { inject_ruffle_and_wait, open_test } = require("../utils");
+const { inject_ruffle_and_wait, open_test } = require("../../utils");
 const { expect, use } = require("chai");
 const chaiHtml = require("chai-html");
 const fs = require("fs");
